+++ sei-db/db_engine/litt/test/keymap_migration_test.go
@@ -31,7 +31,7 @@ func TestKeymapMigration(t *testing.T) {
 	// Build the table using PebbleDBKeymap.
 	config, err := litt.DefaultConfig(shardDirectories...)
 	require.NoError(t, err)
-	config.ShardingFactor = uint32(directoryCount)
+	config.ShardingFactor = uint8(directoryCount)
 	config.KeymapType = keymap.UnsafePebbleDBKeymapType
 	config.Fsync = false // fsync is too slow for unit test workloads
 	config.DoubleWriteProtection = true
@@ -180,7 +180,7 @@ func TestFailedKeymapMigration(t *testing.T) {
 	// Build the table using PebbleDBKeymap.
 	config, err := litt.DefaultConfig(shardDirectories...)
 	require.NoError(t, err)
-	config.ShardingFactor = uint32(directoryCount)
+	config.ShardingFactor = uint8(directoryCount)
 	config.KeymapType = keymap.UnsafePebbleDBKeymapType
 	config.Fsync = false // fsync is too slow for unit test workloads
 	config.DoubleWriteProtection = true
